+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 ## Features
 
 ## Improvements
-
+* `configure_backend` is now exposed to be imported as `from neuroconv.tools import configure_and_write_nwbfile` [#1287](https://github.com/catalystneuro/neuroconv/pull/1287)
 
 # v0.7.2 (April 4, 2025)